+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Addded
 - Array equals method
 - AABB containment and intersection methods
+- `Array.atParam()` method
 
 ### Changed
 - RNG methods are spelled out completely.